feat-addressables
I understood how to manage project resources.
Made a new scene manager.
