+++ Addresables/Assets/AddressableAssetsData/OSX.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b5f1d6ea3ab084b3aabdd7b7d43d9b81
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
