+++ Addresables/Assets/AddressableAssetsData/OSX/addressables_content_state.bin.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5b55c7103d4ae41fab8b8f4f0fac59ea
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
