+++ Addresables/Assets/AddressableAssetsData/ProfileDataSourceSettings.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 742935c3a6aad43328376ea473d6f6e2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
