+++ Addresables/Assets/Project.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9a408b9b142504fb39956d6f8c8edee0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
