+++ Addresables/Assets/Project/Asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6294bd75d04d6498b927da9a25b9626f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
