+++ Addresables/Assets/Project/Scenes/GameScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dc0090744b6ce4f4bb529a38b590e2cb
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
