+++ Addresables/Assets/Project/Scenes/InitScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3af5a1efc6be24fc7bd81edb4456e29b
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
