+++ Addresables/Assets/Project/Scenes/LoadingScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f52df4931e1a24e27886e034a122b84a
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
