+++ Addresables/Assets/Project/Scripts/InitSceneLoader.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+using UnityEngine.AddressableAssets;
+
+namespace Project.Utility
+{
+    public class InitSceneLoader : MonoBehaviour
+    {
+        [SerializeField] private AssetReference _persistentSceneAsset;
+        [SerializeField] private AddressablesManager _addressablesManager;
+
+        private void Start()
+        {
+            Debug.Log("Init scene");
+            _addressablesManager.LoadAddressableScene(_persistentSceneAsset);
+        }
+    }
+}
