+++ Addresables/Assets/Project/Scripts/SceneManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9e37fc21f1524a6cb609e11ad2681ca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
