+++ Addresables/Assets/Project/Scripts/UndestroyableObj.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace Project.Utility
+{
+    public class UndestroyableObj : MonoBehaviour
+    {
+        void Start()
+        {
+            DontDestroyOnLoad(this);
+        }
+    }
+}
