+++ Addresables/Assets/Project/Scripts/UndestroyableObj.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d19761430d98740b1982af51afad7a01
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
