--- Addresables/Assets/Scenes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e7505c4323d824ccf97f20cb71724f18
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
