--- Addresables/Assets/Scenes/InitScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3af5a1efc6be24fc7bd81edb4456e29b
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
