--- Addresables/Assets/Scenes/LoadingScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f52df4931e1a24e27886e034a122b84a
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
