--- Addresables/Assets/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f3cde7839ad134d588a5983acf81110a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
