--- Addresables/Assets/Scripts/AddressablesManager.cs
@@ -1,16 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace Project.Utility
-{
-    public class AddressablesManager : MonoBehaviour
-    {
-        private void Start()
-        {
-            
-        }
-
-
-    }
-}
