--- Addresables/Assets/Scripts/AddressablesManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 42e44e13880ba4763af156ac770e411e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
