--- Addresables/Assets/Scripts/InitSceneLoader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: aeb0ee14eebfd49398d4133a8a9ef458
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
