--- Addresables/Assets/Scripts/UndestroyableObj.cs
@@ -1,12 +1,0 @@
-using UnityEngine;
-
-namespace Project.Utility
-{
-    public class UndestroyableObj : MonoBehaviour
-    {
-        void Start()
-        {
-            DontDestroyOnLoad(this);
-        }
-    }
-}
