+++ Addresables/ProjectSettings/EditorBuildSettings.asset
@@ -6,7 +6,7 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/InitScene.unity
+    path: Assets/Project/Scenes/InitScene.unity
     guid: 3af5a1efc6be24fc7bd81edb4456e29b
   m_configObjects:
     com.unity.addressableassets: {fileID: 11400000, guid: 8acb3142fec0e4bf1b03d1362e6e4278, type: 2}
